refactor: submit handler 공통 로직 훅으로 분리
