+++ Frontend/src/pages/Create.tsx
@@ -1,29 +1,11 @@
-import { FormEvent } from "react";
-
 import Form from "@/components/Form";
-import { FORM_FIELDS } from "@/constants";
 import useCreateBookMutation from "@/hooks/Queries/useCreateBookMutation";
-import { BookType } from "@/types";
+import useFormHandler from "@/hooks/useFormHandler";
 
 const Create = () => {
     const { mutate, isPending } = useCreateBookMutation();
+    const { handleSubmit } = useFormHandler(mutate)
 
-    const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        const formData = new FormData(e.currentTarget);
-
-        const bookData: BookType = FORM_FIELDS.reduce((acc, curField) => {
-            const value = formData.get(curField);
-            if (curField === "quantity") {
-                acc[curField] = value ? Number(value) : 0;
-            } else {
-                acc[curField] = value?.toString() || "";
-            }
-            return acc;
-        }, {} as BookType);
-
-        mutate(bookData);
-    };
     return (
         <Form
             title="도서 등록"
