+++ Frontend/src/pages/Edit.tsx
@@ -1,34 +1,16 @@
-import { FormEvent } from "react";
 import { useParams } from "react-router-dom";
 
 import Form from "@/components/Form";
-import { FORM_FIELDS } from "@/constants";
 import useBookDetailQuery from "@/hooks/Queries/useBookDetailQuery";
 import useEditBookMutation from "@/hooks/Queries/useEditBookMutation";
+import useFormHandler from "@/hooks/useFormHandler";
 import EditFormSkeleton from "@/skeletons/EditFormSkeleton";
-import { BookType } from "@/types";
 
 const Edit = () => {
     const { id } = useParams();
     const { data: bookDetailData, isLoading } = useBookDetailQuery(id!);
     const { mutate, isPending } = useEditBookMutation(Number(id));
-
-    const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-
-        const formData = new FormData(e.currentTarget);
-        const bookData: BookType = FORM_FIELDS.reduce((acc, curField) => {
-            const value = formData.get(curField);
-            if (curField === "quantity") {
-                acc[curField] = value ? Number(value) : 0;
-            } else {
-                acc[curField] = value?.toString() || "";
-            }
-            return acc;
-        }, {} as BookType);
-
-        mutate(bookData);
-    };
+    const { handleSubmit } = useFormHandler(mutate)
 
     if(isLoading) return <EditFormSkeleton/>
